+++ modules-services/prov-service-core/src/main/java/org/openprovenance/prov/service/core/config/EnvironmentVariables.java
@@ -1,0 +1,5 @@
+package org.openprovenance.prov.service.core.config;
+
+public interface EnvironmentVariables {
+    String PROV_STORAGE_CONFIG = "PROV_STORAGE_CONFIG";
+}