+++ modules-services/prov-service-translation/src/main/java/org/openprovenance/prov/service/translation/filesystem/BindingsResourceStorageFileSystem.java
@@ -2,14 +2,13 @@ package org.openprovenance.prov.service.translation.filesystem;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.openprovenance.prov.storage.filesystem.NonDocumentGenericResourceStorageFileSystem;
-import org.openprovenance.prov.template.expander.deprecated.BindingsBean;
+import org.openprovenance.prov.template.json.Bindings;
 
 import java.io.File;
 
-import static org.openprovenance.prov.service.core.ServiceUtils.UPLOADED_FILE_PATH;
 
-public class BindingsResourceStorageFileSystem extends NonDocumentGenericResourceStorageFileSystem<BindingsBean> {
-    public BindingsResourceStorageFileSystem(ObjectMapper om) {
-        super(om, BindingsBean.class, new File(UPLOADED_FILE_PATH));
+public class BindingsResourceStorageFileSystem extends NonDocumentGenericResourceStorageFileSystem<Bindings> {
+    public BindingsResourceStorageFileSystem(ObjectMapper om, File location) {
+        super(om, Bindings.class, location);
     }
 }