+++ modules-services/prov-service-translation/src/main/java/org/openprovenance/prov/service/translation/storage/StorageSetup.java
@@ -6,6 +6,7 @@ import org.apache.logging.log4j.Logger;
 import org.openprovenance.prov.core.xml.serialization.ProvSerialiser;
 import org.openprovenance.prov.model.ProvFactory;
 import org.openprovenance.prov.service.core.ServiceUtilsConfig;
+import org.openprovenance.prov.service.core.config.StorageConfiguration;
 import org.openprovenance.prov.service.core.memory.DocumentResourceIndexInMemory;
 import org.openprovenance.prov.service.translation.actions.ActionExpand;
 import org.openprovenance.prov.service.translation.filesystem.BindingsResourceStorageFileSystem;
@@ -25,109 +26,64 @@ import java.io.OutputStream;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.openprovenance.prov.service.core.ServiceUtils.*;
+public class StorageSetup {
 
-public class StorageConfiguration implements EnvironmentVariables{
+    static Logger logger = LogManager.getLogger(StorageSetup.class);
 
-    static Logger logger = LogManager.getLogger(StorageConfiguration.class);
-
-    public Map<String,String> defaultConfiguration=theDefaultConfiguration();
-
-    public Map<String,String> theDefaultConfiguration() {
-        Map<String,String> config=new HashMap<>();
-        config.put(PSERVICE_DBNAME,     "prov");
-        config.put(PSERVICE_INDEX,      "redis");
-        config.put(PSERVICE_STORAGE,    "mongodb");
-        config.put(PSERVICE_CACHE,      "200");
-        config.put(PSERVICE_AUTODELETE, "true");
-        config.put(PSERVICE_DEL_PERIOD, "1200");
-
-        config.put(PSERVICE_REDIS_HOST, notNullDefault(System.getProperty(REDISSERVER_HOST),"localhost"));
-        config.put(PSERVICE_REDIS_PORT, notNullDefault(System.getProperty(REDISSERVER_PORT), "6379"));
-        config.put(PSERVICE_MONGO_HOST, notNullDefault(System.getProperty(MONGOSERVER_HOST),"localhost"));
-
-        return config;
-    }
-
-
-    static String notNullDefault(String value, String default_value) {
-        System.out.println("value: " + value + ", default: " + default_value);
-        return (value==null)? default_value : value;
-    }
-
-
-    public ServiceUtilsConfig makeConfig(ProvFactory factory) {
-        Map<String,String> config=new HashMap<>();
-        config.putAll(defaultConfiguration);
-        config.putAll(loadConfigFromSystem(defaultConfiguration));
-        config.putAll(loadConfigFromEnvironment(defaultConfiguration));
-
-        return makeConfig(factory,config);
-    }
-
-
-
-    public ServiceUtilsConfig makeConfig(ProvFactory factory, Map<String,String> configuration) {
+    public ServiceUtilsConfig makeConfig(ProvFactory factory, StorageConfiguration configuration) {
         ServiceUtilsConfig utilsConfig = new ServiceUtilsConfig(configuration);
         utilsConfig.pFactory=factory;
         logger.info("Configuration --- " + configuration);
-        switch (configuration.get(PSERVICE_INDEX)) {
+        switch (configuration.index) {
             case "redis":
                 initRedis(utilsConfig, configuration);
                 break;
             case "memory":
                 initInMemory(utilsConfig, configuration);
                 break;
-            default: // do nothing
+            default:
+                throw new IllegalArgumentException("Unknown index type: " + configuration.index);
         }
-        utilsConfig.documentCacheSize = Integer.parseInt(configuration.get(PSERVICE_CACHE));
-        utilsConfig.autoDelete= Boolean.parseBoolean(configuration.get(PSERVICE_AUTODELETE));
-        utilsConfig.deletePeriod= Integer.parseInt(configuration.get(PSERVICE_DEL_PERIOD));
-        switch (configuration.get(PSERVICE_STORAGE)) {
+        utilsConfig.documentCacheSize = configuration.cache;
+        utilsConfig.autoDelete= configuration.autodelete;
+        utilsConfig.deletePeriod= configuration.delete_period;
+        switch (configuration.storage) {
             case "mongodb":
                 withMongoDb(utilsConfig,factory,configuration);
                 break;
             case "fs":
                 withFileSystem(utilsConfig,factory,configuration);
                 break;
-            default: // do nothing
+            default:
+                throw new IllegalArgumentException("Unknown storage type: " + configuration.storage);
         }
-
         return utilsConfig;
     }
 
-    public ServiceUtilsConfig withMongoDb(ServiceUtilsConfig utilsConfig, ProvFactory factory, Map<String,String> configuration) {
+    public ServiceUtilsConfig withMongoDb(ServiceUtilsConfig utilsConfig, ProvFactory factory, StorageConfiguration configuration) {
 
-        final String mongoHost = configuration.get(PSERVICE_MONGO_HOST);
-        final String mongoDbName = configuration.get(PSERVICE_DBNAME);
+        final String mongoHost = configuration.mongo_host;
+        final String mongoDbName = configuration.dbname;
         utilsConfig.storageManager=new MongoDocumentResourceStorage(mongoHost, mongoDbName);
         utilsConfig.nonDocumentResourceStorage=new MongoNonDocumentResourceStorage(mongoHost, mongoDbName);
         utilsConfig.genericResourceStorageMap.put(ActionExpand.BINDINGS_KEY,new MongoBindingsResourceStorage(mongoHost, mongoDbName, new ObjectMapper()));
-
-
-
         ProvSerialiser serial=new ProvSerialiser();
         utilsConfig.serialiser=(OutputStream out, Object document, String mediaType, boolean formatted) -> serial.serialiseOtherObject(out, document,false, false);
-
         return utilsConfig;
     }
 
-    protected File path=new File(UPLOADED_FILE_PATH); //FIXME, integrate with configuration?
 
-
-    public ServiceUtilsConfig withFileSystem(ServiceUtilsConfig utilsConfig, ProvFactory factory, Map<String,String> configuration) {
-        utilsConfig.storageManager=new DocumentResourceStorageFileSystem(factory, path);
-
+    public ServiceUtilsConfig withFileSystem(ServiceUtilsConfig utilsConfig, ProvFactory factory, StorageConfiguration configuration) {
+        File uploadLocation = new File(configuration.uploaded_filepath);
+        utilsConfig.storageManager=new DocumentResourceStorageFileSystem(factory, uploadLocation);
         ProvSerialiser serial = new ProvSerialiser();
         utilsConfig.serialiser=(OutputStream out, Object document, String mediaType, boolean formatted) -> serial.serialiseOtherObject(out, document,false, false);
-
-        utilsConfig.genericResourceStorageMap.put(ActionExpand.BINDINGS_KEY,new BindingsResourceStorageFileSystem(new ObjectMapper()));
-
+        utilsConfig.genericResourceStorageMap.put(ActionExpand.BINDINGS_KEY,new BindingsResourceStorageFileSystem(new ObjectMapper(), uploadLocation));
         return utilsConfig;
     }
 
 
-    public Map<String, ResourceIndex<?>> initInMemory(ServiceUtilsConfig config, Map<String, String> configuration) {
+    public Map<String, ResourceIndex<?>> initInMemory(ServiceUtilsConfig config, StorageConfiguration configuration) {
         Map<String, ResourceIndex<?>> extensionMap = config.extensionMap;
         DocumentResourceIndexInMemory di=new DocumentResourceIndexInMemory();
         extensionMap.put(DocumentResource.getResourceKind(), di);
@@ -136,10 +92,9 @@ public class StorageConfiguration implements EnvironmentVariables{
     }
 
 
-    public Map<String, ResourceIndex<?>> initRedis(ServiceUtilsConfig config, Map<String, String> configuration) {
+    public Map<String, ResourceIndex<?>> initRedis(ServiceUtilsConfig config, StorageConfiguration configuration) {
         Map<String, ResourceIndex<?>> extensionMap = config.extensionMap;
-
-        RedisDocumentResourceIndex di=new RedisDocumentResourceIndex(configuration.get(PSERVICE_REDIS_HOST), Integer.parseInt(configuration.get(PSERVICE_REDIS_PORT)));
+        RedisDocumentResourceIndex di=new RedisDocumentResourceIndex(configuration.redis_host, configuration.redis_port);
         extensionMap.put(DocumentResource.getResourceKind(), di);
         extensionMap.put(TemplateResource.getResourceKind(), new RedisTemplateResourceIndex(di,RedisTemplateResourceIndex.factory));
         return extensionMap;